Integración y refracción completa de ramas, para primera stable release
