+++ lib/components/my_flutter_app_icons.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/widgets.dart';
+
+class MyFlutterApp {
+  MyFlutterApp._();
+
+  static const _kFontFam = 'MyFlutterApp';
+  static const String? _kFontPkg = null;
+
+  static const IconData user =
+      IconData(0xf007, fontFamily: _kFontFam, fontPackage: _kFontPkg);
+  static const IconData trophy =
+      IconData(0xf091, fontFamily: _kFontFam, fontPackage: _kFontPkg);
+  static const IconData chessknight =
+      IconData(0xf441, fontFamily: _kFontFam, fontPackage: _kFontPkg);
+}
