--- lib/my_flutter_app_icons.dart
@@ -1,31 +1,0 @@
-/// Flutter icons MyFlutterApp
-/// Copyright (C) 2023 by original authors @ fluttericon.com, fontello.com
-/// This font was generated by FlutterIcon.com, which is derived from Fontello.
-///
-/// To use this font, place it in your fonts/ directory and include the
-/// following in your pubspec.yaml
-///
-/// flutter:
-///   fonts:
-///    - family:  MyFlutterApp
-///      fonts:
-///       - asset: fonts/MyFlutterApp.ttf
-///
-/// 
-/// * Font Awesome 5, Copyright (C) 2016 by Dave Gandy
-///         Author:    Dave Gandy
-///         License:   SIL (https://github.com/FortAwesome/Font-Awesome/blob/master/LICENSE.txt)
-///         Homepage:  http://fortawesome.github.com/Font-Awesome/
-///
-import 'package:flutter/widgets.dart';
-
-class MyFlutterApp {
-  MyFlutterApp._();
-
-  static const _kFontFam = 'MyFlutterApp';
-  static const String? _kFontPkg = null;
-
-  static const IconData user = IconData(0xf007, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData trophy = IconData(0xf091, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-  static const IconData chess_knight = IconData(0xf441, fontFamily: _kFontFam, fontPackage: _kFontPkg);
-}
