+++ lib/pages/menus_pages/profile.dart
@@ -10,7 +10,7 @@ class ProfileScreen extends StatefulWidget {
 class _ProfileScreenState extends State<ProfileScreen> {
   @override
   Widget build(BuildContext context) {
-    return Text(
+    return const Text(
       'Profile',
     );
   }
