+++ pubspec.yaml
@@ -63,10 +63,6 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - images/
-    - lib/images/logo.png
-    - lib/images/caballo.png
-    - lib/images/diana.png
-    - lib/images/trofeo.png
     
 
   # An image asset can refer to one or more resolution-specific "variants", see
@@ -83,7 +79,7 @@ flutter:
   fonts:
      - family:  MyFlutterApp
        fonts:
-         - asset: lib/fonts/MyFlutterApp.ttf
+         - asset: fonts/MyFlutterApp.ttf
      - family: VesperLibre
        fonts:
          - asset: fonts/VesperLibre-Black.ttf
